Revert "refactor: render style setter"
