+++ integration_tests/specs/css/css-display/sliver.ts
@@ -277,87 +277,4 @@ describe('display sliver', () => {
     container.childNodes[1].appendChild(absoluted);
     await snapshot();
   });
-
-  it('should work with child style has been maintained when child is destroyed and recreated', async (done) => {
-    let container = createElement('div', {
-      style: {
-        display: 'sliver',
-        width: '200px',
-        height: '180px',
-      }
-    });
-    let div0;
-
-    for (let i = 0; i < 5; i++) {
-      let div = createElement('div', {
-        style: {
-          border: '1px solid green',
-          backgroundColor: 'yellow',
-          color: 'red',
-          opacity: '0.5',
-          width: '200px',
-          height: '100px',
-          overflowX: 'scroll',
-          whiteSpace: 'nowrap',
-          position: 'relative'
-        } 
-      }, [
-        createText(i + ' aaaaaa bbbbbb cccccc dddddd eeeeee ffffff gggggg'),
-        createElement('div', {
-          style: {
-            margin: '20px 0 0 20px',
-            width: '100px',
-            height: '50px',
-            backgroundColor: 'coral'
-          }
-        }, [
-          createElement('div', {
-            style: {
-              position: 'absolute',
-              left: '30px',
-              top: '20px',
-              width: '50px',
-              height: '50px',
-              backgroundColor: 'green',
-              zIndex: 2
-            }
-          }),
-          createElement('div', {
-            style: {
-              position: 'sticky',
-              left: '50px',
-              width: '50px',
-              height: '50px',
-              backgroundColor: 'blue',
-              zIndex: 1
-            }
-          })
-        ])
-      ]);
-      if (i == 0) {
-        div0 = div;
-      }
-
-      container.appendChild(div);
-    }
-
-    BODY.appendChild(container);
-
-    await snapshot();
-
-    requestAnimationFrame(async () => {
-      div0.scrollTo(1000, 0);
-      await snapshot();
-
-      container.scrollTo(0, 1000);
-      await snapshot();
-
-      container.scrollTo(0, -1000);
-      await snapshot();
-
-      div0.scrollTo(-1000, 0);
-      await snapshot();
-      done();
-    });
-  });
 });
