+++ kraken/lib/dom.dart
@@ -4,7 +4,6 @@
 
 export 'src/dom/binding.dart';
 export 'src/dom/element.dart';
-export 'src/dom/overflow.dart';
 export 'src/dom/event.dart';
 export 'src/dom/event_target.dart';
 export 'src/dom/object_element_client.dart';
