+++ kraken/lib/src/rendering/opacity.dart
@@ -1,14 +1,14 @@
 /*
  * Copyright (C) 2019-present The Kraken authors. All rights reserved.
  */
-import 'package:flutter/rendering.dart';
-import 'package:kraken/rendering.dart';
+import 'dart:ui' as ui;
 
-mixin RenderOpacityMixin on RenderBoxModelBase {
-  bool opacityAlwaysNeedsCompositing() {
-    int alpha = renderStyle.alpha;
-    return alpha != 0 && alpha != 255;
-  }
+import 'package:flutter/rendering.dart';
+
+mixin RenderOpacityMixin on RenderBox {
+  bool opacityAlwaysNeedsCompositing() => alpha != 0 && alpha != 255;
+
+  int alpha = ui.Color.getAlphaFromOpacity(1.0);
 
   final LayerHandle<OpacityLayer> _opacityLayer = LayerHandle<OpacityLayer>();
 
@@ -18,7 +18,6 @@ mixin RenderOpacityMixin on RenderBoxModelBase {
 
   void paintOpacity(PaintingContext context, Offset offset,
       PaintingContextCallback callback) {
-    int alpha = renderStyle.alpha;
 
     if (alpha == 255) {
       _opacityLayer.layer = null;
@@ -32,7 +31,6 @@ mixin RenderOpacityMixin on RenderBoxModelBase {
   }
 
   void debugOpacityProperties(DiagnosticPropertiesBuilder properties) {
-    int alpha = renderStyle.alpha;
     if (alpha != 0 && alpha != 255)
       properties.add(DiagnosticsProperty('alpha', alpha));
   }
