+++ kraken/lib/src/rendering/sliver_list.dart
@@ -43,17 +43,14 @@ class RenderSliverListLayout extends RenderLayoutBox {
     scrollable = KrakenScrollable(axisDirection: getAxisDirection(axis));
     axis = renderStyle.sliverDirection;
 
-    renderStyle.target.scrollablePointerListener = _scrollablePointerListener;
-    renderStyle.target.scrollListener = _scrollListener;
-
     switch (axis) {
       case Axis.horizontal:
-        renderStyle.target.scrollOffsetX = scrollable.position;
-        renderStyle.target.scrollOffsetY = null;
+        scrollOffsetX = scrollable.position;
+        scrollOffsetY = null;
         break;
       case Axis.vertical:
-        renderStyle.target.scrollOffsetX = null;
-        renderStyle.target.scrollOffsetY = scrollable.position;
+        scrollOffsetX = null;
+        scrollOffsetY = scrollable.position;
         break;
     }
 
@@ -67,6 +64,13 @@ class RenderSliverListLayout extends RenderLayoutBox {
     manager.setupSliverListLayout(this);
     super.insert(_renderViewport);
   }
+
+  // Override the scrollable pointer listener.
+  @override
+  void Function(PointerEvent event) get scrollablePointerListener => _scrollablePointerListener;
+
+  @override
+  ScrollListener? get scrollListener => _scrollListener;
 
   @override
   bool get isRepaintBoundary => true;
